+++ build.clj
@@ -8,7 +8,7 @@
 
 (defonce project
   (let [lib        'com.tolitius/obiwan
-        version    "0.2.4"
+        version    "0.2.5"
         target-dir "target"]
     {:lib lib
      :description "redis clojure client based on jedis"